+++ playwright/tests/sso_groups_groups.spec.ts
@@ -22,7 +22,6 @@ test.beforeAll('Setup', async ({ browser }, testInfo: TestInfo) => {
         SSO_ENABLED: true,
         SSO_ONLY: true,
         SSO_ORGANIZATIONS_ENABLED: true,
-        SSO_ORGANIZATIONS_GROUPS_ENABLED: true,
         SSO_ORGANIZATIONS_REVOCATION: true,
         SSO_SCOPES: "email profile groups",
         SMTP_HOST: process.env.MAILDEV_HOST,